change type

+++ autos/models.py
@@ -20,5 +20,5 @@ class Position(models.Model):
     latitude = models.DecimalField(max_digits=20, decimal_places=10)
     longitude = models.DecimalField(max_digits=20, decimal_places=10)
     date_time = models.DateTimeField(null=True)
-    speed = models.IntegerField(default=0)
+    speed = models.FloatField(default=0)
     distance = models.FloatField(default=0)
